BIMoph working MVVM
MVVM wotking
