+++ Purge Rooms UI/packages.config
@@ -1,6 +1,8 @@
 ﻿<?xml version="1.0" encoding="utf-8"?>
 <packages>
   <package id="CommonServiceLocator" version="2.0.2" targetFramework="net481" />
+  <package id="MvvmLight" version="5.4.1.1" targetFramework="net48" />
   <package id="MvvmLightLibs" version="5.4.1.1" targetFramework="net48" />
   <package id="Revit_All_Main_Versions_API_x64" version="2022.1.0" targetFramework="net48" />
+  <package id="System.ValueTuple" version="4.5.0" targetFramework="net48" />
 </packages>
